update file structures
