+++ apps/web/src/app/auth/login/page.tsx
@@ -5,7 +5,6 @@ import Link from 'next/link';
 import { useForm } from 'react-hook-form';
 
 // import type { Metadata } from 'next';
-
 import {
   Form,
   FormControl,
@@ -16,26 +15,25 @@ import {
 } from '~/components/form';
 import { Button } from '~/components/ui/button';
 import { Input } from '~/components/ui/input';
-import { registerSchema, type RegisterSchema } from '~/schemas/index.schema';
+import { loginSchema, type LoginSchema } from '~/schemas/index.schema';
 
 // const metadata: Metadata = {
-//   title: 'Login',
-//   description: 'Login page for the app.',
+//  title: 'Login',
+//  description: 'Login page for the app.',
 // };
 
 export default function LoginPage() {
   // 1. Define your form.
-  const form = useForm<RegisterSchema>({
-    resolver: zodResolver(registerSchema),
+  const form = useForm<LoginSchema>({
+    resolver: zodResolver(loginSchema),
     defaultValues: {
-      username: '',
       email: '',
       password: '',
     },
   });
 
   // 2. Define a submit handler.
-  function onSubmit(values: RegisterSchema) {
+  function onSubmit(values: LoginSchema) {
     // Do something with the form values.
     // ✅ This will be type-safe and validated.
     console.log(values);
@@ -47,31 +45,14 @@ export default function LoginPage() {
       <div className="flex w-full flex-col justify-center space-y-6 sm:w-[400px] border border-slate-500 dark:border-slate-700 p-8 rounded-lg">
         <div className="flex flex-col space-y-2 text-center">
           <h1 className="text-2xl font-semibold tracking-tight">
-            Create an account
+            Login to your account
           </h1>
           <p className="text-muted-foreground text-sm">
-            Register an account to start using our service
+            Enter your credentials below to log in into your account
           </p>
         </div>
         <Form {...form}>
           <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6">
-            <FormField
-              control={form.control}
-              name="username"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Username</FormLabel>
-                  <FormControl>
-                    <Input
-                      className="border-slate-500 dark:border-slate-700"
-                      placeholder="johndoe"
-                      {...field}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
             <FormField
               control={form.control}
               name="email"
@@ -107,7 +88,7 @@ export default function LoginPage() {
               )}
             />
             <Button type="submit" className="w-full">
-              Create an account
+              Login
             </Button>
           </form>
         </Form>
@@ -129,12 +110,12 @@ export default function LoginPage() {
           .
         </p>
         <p className="text-muted-foreground flex justify-between items-center">
-          Already have an account?
+          Don&apos;t have an account?{' '}
           <Link
-            href="/login"
+            href="/auth/register"
             className="hover:text-primary underline underline-offset-4"
           >
-            login here
+            register here
           </Link>
         </p>
       </div>
